Add files via upload
"Advertising is full of change that requires immediate responses to client needs. We are fortunately growing and need additional creatives to help service our clients. We are looking for a freelance frontend developer that is passionate about brand growth through engaging digital experiences, has experience working with agencies, and can plug-n-play with our brand teams on various campaigns and projects.

In most cases, our frontend developers are directly involved with the creative process working alongside our copywriter, project manager, digital marketer, and art director. We also welcome our freelancers to work with us in our office or workspace locations. While it often isn’t required, in many cases we encourage our freelancers to attend client meetings for collaboration.

We are looking for a frontend developer that possesses a strong sense of web and mobile development for WordPress. The ideal person would also be strong on responsive, fluid, or adaptive executions that work perfectly on all mobile devices.

Ideally, we’d like to work with freelancers for the long haul because we believe that relationships are the key to creative excellence. We are also looking to bring fresh execution in the following domains: js, css, speed optimization, animations, and transitions. Currently, we are looking to fill in positions and balance the workload on brands that include a Top 100 Roofing Company, nationally-ranked Division I School, State of Texas, and our own agency.

If you’re passionate about creating and curating ideas that are brought to life within campaign execution, please apply. We’d love to hear from you.

Visit Us: https://recruiting.foundry512.com/job/frontend-developer/

and the agency site: https://www.foundry512.com/"